feat: throw explicit error if smtp is not configured
