+++ apps/web/app/(app)/environments/[environmentId]/settings/members/page.tsx
@@ -28,8 +28,10 @@ const MembersLoading = () => (
     </div>
 
     <div className="p-4">
-      {[1, 2, 3].map((_) => (
-        <div className="grid-cols-20 grid h-12 content-center rounded-t-lg bg-white p-4 text-left text-sm font-semibold text-slate-900">
+      {[1, 2, 3].map((i) => (
+        <div
+          key={i}
+          className="grid-cols-20 grid h-12 content-center rounded-t-lg bg-white p-4 text-left text-sm font-semibold text-slate-900">
           <Skeleton className="col-span-2 h-10 w-10 rounded-full" />
           <Skeleton className="col-span-5 h-8 w-24" />
           <Skeleton className="col-span-5 h-8 w-24" />
