+++ apps/web/app/api/v1/users/forgot-password/route.ts
@@ -1,7 +1,7 @@
-import { sendForgotPasswordEmail } from "@/app/lib/email";
 import { NextResponse } from "next/server";
 
 import { prisma } from "@formbricks/database";
+import { sendForgotPasswordEmail } from "@formbricks/lib/emails/emails";
 
 export async function POST(request: Request) {
   const { email } = await request.json();
